работа с сетью + рефакторинг
