+++ src/view/event-edit/template.js
@@ -1,11 +1,7 @@
-import { BLANK_TRIP_EVENT, EVENT_TYPES, DateFormats, ButtonTypes, DefaultFlatpickrConfig } from '../const';
-import AbstractStatefulView from '../framework/view/abstract-stateful-view';
-import { displayDateTime } from '../utils/date';
-import { remove } from '../framework/render';
-import { firstLetterUpperCase, getIsCheckedAttr, getIsDisabledAttr, getInteger, addItem, removeItem } from '../utils/common';
+import { EVENT_TYPES, DateFormats, ButtonTypes } from '../../const';
+import { displayDateTime } from '../../utils/date';
+import { firstLetterUpperCase, getIsCheckedAttr, getIsDisabledAttr } from '../../utils/common';
 import he from 'he';
-import flatpickr from 'flatpickr';
-import 'flatpickr/dist/flatpickr.min.css';
 
 const getTypeItemTemplate = (type, isChecked) => `
   <div class="event__type-item">
@@ -101,7 +97,7 @@ const getOffersTemplate = (offers) => !offers.length ? '' : `
     </div>
   </section>`;
 
-const createPhotoTapeTemplate = (pictures) => !pictures.length ? '' : `
+const getPhotoTapeTemplate = (pictures) => !pictures.length ? '' : `
   <div class="event__photos-container">
     <div class="event__photos-tape">
       ${pictures.map(({src, description}) => `<img class="event__photo" src="${src}" alt="${description}">`).join('')}
@@ -117,11 +113,11 @@ const getDestinationTemplate = (destination) => {
     <section class="event__section  event__section--destination">
       <h3 class="event__section-title  event__section-title--destination">Destination</h3>
       <p class="event__destination-description">${description}</p>
-      ${createPhotoTapeTemplate(pictures)}
+      ${getPhotoTapeTemplate(pictures)}
     </section>`;
 };
 
-const createEventEditTemplate = (tripEvent, offers, destinations) => {
+const getEventEditTemplate = (tripEvent, offers, destinations) => {
   const { type, dateFrom, dateTo, basePrice, isAdding, isSaving, isDeleting } = tripEvent;
   const eventDestination = destinations.find((destination) => destination.id === tripEvent.destination);
   const { offers: typedOffers } = offers.find((offer) => offer.type === type);
@@ -147,177 +143,4 @@ const createEventEditTemplate = (tripEvent, offers, destinations) => {
   </form>`;
 };
 
-export default class EventEditView extends AbstractStatefulView {
-  #offers = null;
-  #destinations = null;
-  #submitHandler = null;
-  #deleteHandler = null;
-  #cancelHandler = null;
-  #dateFromPicker = null;
-  #dateToPicker = null;
-
-  constructor({tripEvent = BLANK_TRIP_EVENT, offers, destinations, onFormSubmit, onFormDelete, onFormCancel}) {
-    super();
-    this._setState(EventEditView.parseEventToState(tripEvent));
-    this.#offers = offers;
-    this.#destinations = destinations;
-    this.#submitHandler = onFormSubmit;
-    this.#deleteHandler = onFormDelete;
-    this.#cancelHandler = onFormCancel;
-
-    this._restoreHandlers();
-  }
-
-  get template() {
-    return createEventEditTemplate(this._state, this.#offers, this.#destinations);
-  }
-
-  _restoreHandlers() {
-    this.element.addEventListener('submit', this.#onFormSubmit);
-    this.element.querySelector('.event__reset-btn').addEventListener('click', this.#getResetHandler());
-    this.element.querySelector('.event__type-group').addEventListener('change', this.#onTypeChange);
-    this.element.querySelector('.event__input--destination').addEventListener('change', this.#onDestinationChange);
-    this.element.querySelector('.event__input--price').addEventListener('change', this.#onPriceChange);
-    this.element.querySelector('.event__input--price').addEventListener('input', this.#onPriceInput);
-
-    const availableOffers = this.element.querySelector('.event__available-offers');
-    if (availableOffers) {
-      availableOffers.addEventListener('change', this.#onOfferClick);
-    }
-    const rollupButtonElement = this.element.querySelector('.event__rollup-btn');
-    if (rollupButtonElement) {
-      rollupButtonElement.addEventListener('click', this.#onCancelForm);
-    }
-
-    this.#setDatePickers({
-      startTimeElement: this.element.querySelector('#event-start-time-1'),
-      endTimeElement: this.element.querySelector('#event-end-time-1')
-    });
-  }
-
-  reset(tripEvent) {
-    this.updateElement(tripEvent);
-  }
-
-  destroy() {
-    remove(this);
-  }
-
-  removeElement() {
-    this.#dateFromPicker.destroy();
-    this.#dateFromPicker = null;
-    this.#dateToPicker.destroy();
-    this.#dateToPicker = null;
-
-    super.removeElement();
-  }
-
-  #getResetHandler = () => this._state.isAdding ? this.#onCancelForm : this.#onDeleteForm;
-
-  #setDatePickers({ startTimeElement, endTimeElement }) {
-    this.#dateFromPicker = flatpickr(
-      startTimeElement,
-      {
-        ...DefaultFlatpickrConfig,
-        defaultDate: this._state.dateFrom,
-        maxDate: this._state.dateTo,
-        onChange: this.#onDateFromChange,
-      },
-    );
-
-    this.#dateToPicker = flatpickr(
-      endTimeElement,
-      {
-        ...DefaultFlatpickrConfig,
-        defaultDate: this._state.dateTo,
-        minDate: this._state.dateFrom,
-        onChange: this.#onDateToChange,
-      },
-    );
-  }
-
-  #onFormSubmit = (evt) => {
-    evt.preventDefault();
-    this.#submitHandler(EventEditView.parseStateToEvent(this._state));
-  };
-
-  #onDeleteForm = (evt) => {
-    evt.preventDefault();
-    this.#deleteHandler(EventEditView.parseStateToEvent(this._state));
-  };
-
-  #onCancelForm = (evt) => {
-    evt.preventDefault();
-    this.#cancelHandler();
-  };
-
-  #onTypeChange = (evt) => {
-    const type = evt.target.value;
-    if (this._state.type === type) {
-      return;
-    }
-    this.updateElement({
-      type,
-      offers: [],
-    });
-  };
-
-  #onDestinationChange = (evt) => {
-    const destination = this.#destinations.find(({ name }) => name === evt.target.value);
-    if (!destination || this._state.destination === destination.id) {
-      return;
-    }
-    this.updateElement({
-      destination: destination.id,
-    });
-  };
-
-  #onPriceInput = (evt) => {
-    evt.target.value = getInteger(evt.target.value);
-  };
-
-  #onPriceChange = (evt) => {
-    const price = getInteger(evt.target.value);
-    this.updateElement({
-      basePrice: price,
-    });
-  };
-
-  #onDateFromChange = ([date]) => {
-    this.updateElement({
-      dateFrom: date,
-    });
-  };
-
-  #onDateToChange = ([date]) => {
-    this.updateElement({
-      dateTo: date,
-    });
-  };
-
-  #onOfferClick = (evt) => {
-    const { dataset: { offerId }, checked } = evt.target;
-    const offers = checked
-      ? addItem(this._state.offers, offerId)
-      : removeItem(this._state.offers, offerId);
-
-    this.updateElement({
-      offers,
-    });
-  };
-
-  static parseEventToState = (tripEvent) => ({
-    ...tripEvent,
-    isAdding: !tripEvent.id,
-    isSaving: false,
-    isDeleting: false,
-  });
-
-  static parseStateToEvent = (state) => {
-    const tripEvent = {...state};
-    delete tripEvent.isAdding;
-    delete tripEvent.isSaving;
-    delete tripEvent.isDeleting;
-    return tripEvent;
-  };
-}
+export { getEventEditTemplate };
