+++ src/view/trip-filters-view.js
@@ -37,14 +37,12 @@ export default class TripFiltersView extends AbstractView {
     return createFiltersTemplate(this.#filters, this.#currentFilter);
   }
 
-  destroy() {
-    remove(this);
-  }
+  destroy = () => remove(this);
 
-  removeElement() {
+  removeElement = () => {
     this.element.removeEventListener('change', this.#onFilterChange);
     super.removeElement();
-  }
+  };
 
   #onFilterChange = (evt) => {
     evt.preventDefault();
